settings: make RegisterEnumSetting generic
We change `RegisterEnumSetting` to use any integer type instead of
always `int64`. This makes the code safer and cleaner (we no longer
need to/from `int64` casts).

Epic: none
Release note: None

+++ pkg/kv/kvserver/allocator/allocatorimpl/allocator_test.go
@@ -2124,7 +2124,7 @@ func TestAllocatorTransferLeaseTargetIOOverloadCheck(t *testing.T) {
 
 			sg := gossiputil.NewStoreGossiper(g)
 			sg.GossipStores(stores, t)
-			LeaseIOOverloadThresholdEnforcement.Override(ctx, &a.st.SV, int64(tc.enforcement))
+			LeaseIOOverloadThresholdEnforcement.Override(ctx, &a.st.SV, tc.enforcement)
 			LeaseIOOverloadThreshold.Override(ctx, &a.st.SV, threshold)
 			LeaseIOOverloadShedThreshold.Override(ctx, &a.st.SV, shedThreshold)
 
@@ -2955,7 +2955,7 @@ func TestAllocatorShouldTransferLeaseIOOverload(t *testing.T) {
 
 			sg := gossiputil.NewStoreGossiper(g)
 			sg.GossipStores(stores, t)
-			LeaseIOOverloadThresholdEnforcement.Override(ctx, &a.st.SV, int64(tc.enforcement))
+			LeaseIOOverloadThresholdEnforcement.Override(ctx, &a.st.SV, tc.enforcement)
 			LeaseIOOverloadThreshold.Override(ctx, &a.st.SV, threshold)
 			LeaseIOOverloadShedThreshold.Override(ctx, &a.st.SV, shedThreshold)
