+++ pkg/kv/kvserver/kvflowcontrol/kvflowhandle/kvflowhandle_test.go
@@ -89,7 +89,7 @@ func TestHandleAdmit(t *testing.T) {
 			clock := hlc.NewClockForTesting(nil)
 			st := cluster.MakeTestingClusterSettings()
 			kvflowcontrol.Enabled.Override(ctx, &st.SV, true)
-			kvflowcontrol.Mode.Override(ctx, &st.SV, int64(kvflowcontrol.ApplyToAll))
+			kvflowcontrol.Mode.Override(ctx, &st.SV, kvflowcontrol.ApplyToAll)
 
 			controller := kvflowcontroller.New(registry, st, clock)
 			handle := kvflowhandle.New(
@@ -173,7 +173,7 @@ func TestFlowControlMode(t *testing.T) {
 			clock := hlc.NewClockForTesting(nil)
 			st := cluster.MakeTestingClusterSettings()
 			kvflowcontrol.Enabled.Override(ctx, &st.SV, true)
-			kvflowcontrol.Mode.Override(ctx, &st.SV, int64(tc.mode))
+			kvflowcontrol.Mode.Override(ctx, &st.SV, tc.mode)
 
 			controller := kvflowcontroller.New(registry, st, clock)
 			handle := kvflowhandle.New(
@@ -253,7 +253,7 @@ func TestInspectHandle(t *testing.T) {
 	clock := hlc.NewClockForTesting(nil)
 	st := cluster.MakeTestingClusterSettings()
 	kvflowcontrol.Enabled.Override(ctx, &st.SV, true)
-	kvflowcontrol.Mode.Override(ctx, &st.SV, int64(kvflowcontrol.ApplyToAll))
+	kvflowcontrol.Mode.Override(ctx, &st.SV, kvflowcontrol.ApplyToAll)
 
 	pos := func(d uint64) kvflowcontrolpb.RaftLogPosition {
 		return kvflowcontrolpb.RaftLogPosition{Term: 1, Index: d}
